requirements-dev.txt
ruff
mypy
pytest
pytest-asyncio
pytest-cov
pandas-stubs
matplotlib